requirements.txt
facebook-business==19.0.0
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
requests==2.31.0
